third_party/WebKit/LayoutTests/editing/deleting/delete-mixed-editable-content-002-expected.txt

Test case for bug http://crbug.com/238000 Forward and backward delete nested content editable

Dump of markup 1:
| "
  "
| <span>
|   "Start<#selection-caret>"
| <span>
|   "(delete)"
|   "End"
| "
"

Dump of markup 2:
| "
  "
| <span>
|   "Start"
|   "(forwardDelete)<#selection-caret>"
| <span>
|   "End"
| "
"
